Update repositories.txt remove EdgieD forever
The library submission process makes no sense to me at all. I quit. This is one of the reasons I moved to PlatformIO, adding libraries is on a project-by-project basis, where you simply add a link to a git repo in the project's platformio.ini file. Even version 2 of Arduino IDE is outdated rubbish.