Remove "fa3el_khar"
The maintainer of this library has established a pattern of disrespectful and irresponsible behavior in the
arduino/library-registry repository despite repeated requests to desist. For this reason, Arduino is revoking their
privileges of distributing libraries via the Library Manager system.

The library is an identical copy of "ArduinoMDNS", except with the addition of a trivial example sketch that doesn't
even use the library or have any relation to the subject of MDNS, and the replacement of the readme text with "LOL". So
it is clear this is not a good faith attempt to make a true contribution to the Library Manager and the removal will
only benefit the Arduino community.