Replace broken Arduino Cloud Library Manager reference link
The FAQ contains an introduction to the Arduino Library Manager. This is supplemented with a list of links to references
about the Library Manager interface in each of the Arduino development tools.

The never ending churn of the Arduino IDE website structure resulted in the breakage of the link to the Arduino Cloud
reference. It is hereby updated to point to the replacement page.