Use standard URL format for "K24C16 EEPROM Library"
The established convention is to use the standard web URL format for the entries in the repositories.txt file.

Previously, the entry for the "K24C16 EEPROM Library" used the alternative URL format with the `.git` extension. The
Library Manage registry system is designed to support either one, so that was not a problem from a technical standpoint,
but it is better to use a consistent format throughout the file.