Update reference URL for github.com/galihru privilege revocation
The previous URL was for the original privilege revocation. Since they chose to continue their irresponsible behavior
instead of making a privilege restoration request as was detailed in the original revocation notification, the
revocation is now permanent. Thus it does not make sense to link to the original notification which contains a
restoration offer.