Remove github.com/VasilKalchev/RGBLED registration
It seems that a compliant release of this library was never created, meaning that although there is a registry entry,
the library is not actually listed in the libraries index.

The registration occurred eight years ago and the repository has since either been deleted or made private, so a release
of the library will never come. Thus, removing the registration will not cause any harm.

The existence of this "ghost registration" is problematic because the duplicate name check system uses the libraries
index. Since the library is not present there, the check doesn't see that the name "RGBLED" has already been registered
and thus will allow registration of additional libraries with that name. In addition, its presence may cause confusion
to the registry maintainers since we expect each registration to have a unique name.

Since the registration of the library provides no benefit and causes harm, it is removed.