Trim trailing whitespace from repositories list
Trailing whitespace was introduced by a pull request. Some editors automatically trim trailing whitespace. If a
contributor uses such an editor when making a submission, the resulting diff will cause the automated system to detect
the pull request as being of the "modification" type instead of the "submission" type. This will cause the pull request
to be incorrectly handled.